+++ src/antity.ts
@@ -1,5 +1,5 @@
 import { isArray, isObject, isString, isIn } from '@dwtechs/checkard';
-// import { log } from "@dwtechs/winstan";
+import { log } from "@dwtechs/winstan";
 import { Property } from './property';
 import { Messages } from './message';
 import { Types, Required } from './checks';
@@ -78,7 +78,7 @@ export class Entity {
     for (const r of rows) {
       for (const { 
         key, 
-        // type,
+        type,
         sanitize,
         normalize,
         sanitizer,
@@ -87,11 +87,11 @@ export class Entity {
         let v = r[key];
         if (v) {
           if (sanitize) {
-            // log.debug(`sanitize ${key}: ${type} = ${v}`);
+            log.debug(`sanitize ${key}: ${type} = ${v}`);
             v = this.sanitize(v, sanitizer);
           }
           if (normalize && normalizer) {
-            // log.debug(`normalize ${key}: ${type} = ${v}`);
+            log.debug(`normalize ${key}: ${type} = ${v}`);
             v = normalizer(v);
           }
           r[key] = v;
@@ -120,7 +120,7 @@ export class Entity {
         const v = r[key];
         if (isIn(operation, operations)) {
           if (required) {
-            const rq = this.require(v, key/*, type*/);
+            const rq = this.require(v, key, type);
             if (rq)
               return rq;
           }
@@ -135,8 +135,8 @@ export class Entity {
     return null;
   }
     
-  private require(v: any, key: string/*, type: Type*/): any {
-    // log.debug(`require ${key}: ${type} = ${v}`);	
+  private require(v: any, key: string, type: Type): any {
+    log.debug(`require ${key}: ${type} = ${v}`);	
     return Required.validate(v) ? null : Messages.missing(key);
   }
 
@@ -149,7 +149,7 @@ export class Entity {
     typeCheck: boolean,
     cb: ((v:any) => any) | null
   ): any {
-    // log.debug(`control ${key}: ${type} = ${v}`);
+    log.debug(`control ${key}: ${type} = ${v}`);
     if (cb)
       return cb(v) ? null : Messages.invalid(key, type);
     return Types[type].validate(v, min, max, typeCheck) ? null : Messages.invalid(key, type);