+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - add getUnsafeProps() method to get an unsafe properties array
 - replace REST methods by SQL operations
 - delete "name" property
-- add pagination property to getCols() method to add a total count in a select query  
+- add stringify property to getCols() method to tell whether to return columns as string or array.
+- add pagination property to getCols() method to add a total count in a select query 
+- add logs using @dwtechs/winstan library
 
 # 0.3.0 (Dec 30th 2024)