+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - add getUnsafeProps() method to get an unsafe properties array
 - update operation parameter possible values to the validate() method with SQL operations and REST methods to validate proper properties depending on the current action 
 - delete "name" property
-- add stringify property to getCols() method to tell whether to return columns as string or array.
-- add pagination property to getCols() method to add a total count in a select query 
+- add stringify property to getCols() method to tell whether to return columns as string or array
+- add pagination property to getCols() method to add a total count in a select query
 
 # 0.3.0 (Dec 30th 2024)